server/requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
psycopg2-binary==2.9.11
sqlalchemy==1.4.31
pydantic==1.9.0
python-dotenv==0.19.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pyjwt==2.3.0
python-jose==3.5.0
